[FEATURE] Get Orders

+++ src/main/java/com/umeal/api/order/repository/OrderRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 @Repository
 public interface OrderRepository extends JpaRepository<Order, Long> {
-    List<Order> findByUserId(Long userId);
+    List<Order> findByUserId(Integer userId);
 
     List<Order> findByRestaurantId(Long restaurantId);
 }
